Refactor ContainerChainSpawner into separate module
