+++ node/src/service.rs
@@ -1,9 +1,7 @@
 //! Service and ServiceFactory implementation. Specialized wrapper over substrate service.
-use exit_future::Signal;
-use futures::FutureExt;
-use sc_service::SpawnTaskHandle;
-use std::collections::HashMap;
-use std::pin::Pin;
+use crate::container_chain_spawner::CcSpawnMsg;
+use crate::container_chain_spawner::ContainerChainSpawner;
+use crate::container_chain_spawner::CCSPAWN;
 use tokio::sync::mpsc::{unbounded_channel, UnboundedSender};
 use {
     crate::cli::ContainerChainCli,
@@ -47,7 +45,7 @@ use {
     sp_consensus::SyncOracle,
     sp_keystore::{SyncCryptoStore, SyncCryptoStorePtr},
     sp_state_machine::{Backend as StateBackend, StorageValue},
-    std::{future::Future, str::FromStr, sync::Arc, sync::Mutex, time::Duration},
+    std::{str::FromStr, sync::Arc, sync::Mutex, time::Duration},
     substrate_prometheus_endpoint::Registry,
     tc_consensus::{BuildOrchestratorAuraConsensusParams, OrchestratorAuraConsensus},
     tc_orchestrator_chain_interface::{
@@ -76,7 +74,7 @@ impl sc_executor::NativeExecutionDispatch for ParachainNativeExecutor {
 
 type ParachainExecutor = NativeElseWasmExecutor<ParachainNativeExecutor>;
 
-type ParachainClient = TFullClient<Block, RuntimeApi, ParachainExecutor>;
+pub type ParachainClient = TFullClient<Block, RuntimeApi, ParachainExecutor>;
 
 type ParachainBackend = TFullBackend<Block>;
 
@@ -583,281 +581,6 @@ async fn start_node_impl(
     Ok((task_manager, client))
 }
 
-/// Struct with all the params needed to start a container chain node given the CLI arguments,
-/// and creating the ChainSpec from on-chain data from the orchestrator chain.
-#[derive(Clone)]
-struct ContainerChainSpawner {
-    // Start container chain params
-    orchestrator_chain_interface: Arc<dyn OrchestratorChainInterface>,
-    orchestrator_client: Arc<ParachainClient>,
-    container_chain_cli: ContainerChainCli,
-    tokio_handle: tokio::runtime::Handle,
-    chain_type: sc_chain_spec::ChainType,
-    relay_chain: String,
-    relay_chain_interface: Arc<dyn RelayChainInterface>,
-    collator_key: Option<CollatorPair>,
-    sync_keystore: SyncCryptoStorePtr,
-    orchestrator_para_id: ParaId,
-    validator: bool,
-    spawn_handle: SpawnTaskHandle,
-
-    // State
-    spawned_para_ids: Arc<Mutex<HashMap<ParaId, StopContainerChain>>>,
-
-    // Collate on Tanssi
-    collate_on_tanssi: Arc<dyn Fn() -> Pin<Box<dyn Future<Output = ()> + Send>> + Send + Sync>,
-}
-
-struct StopContainerChain(Signal);
-
-/// Messages used to control the `ContainerChainSpawner`. This is needed because one of the fields
-/// of `ContainerChainSpawner` is not `Sync`, so we cannot simply pass an
-/// `Arc<ContainerChainSpawner>` to other threads.
-#[derive(Debug)]
-pub enum CcSpawnMsg {
-    /// Start a container chain client for this ParaId. If the ParaId is the orchestrator chain id,
-    /// start collating there.
-    Spawn(ParaId),
-    /// Stop the container chain client previously started for this ParaId. If the ParaId is the
-    /// orchestrator chain id, ignore this message.
-    Stop(ParaId),
-}
-
-// TODO: this is for testing, remove when not needed and remove lazy_static dependency
-lazy_static::lazy_static! {
-    static ref CCSPAWN: Mutex<Option<UnboundedSender<CcSpawnMsg>>> = Mutex::new(None);
-}
-
-impl ContainerChainSpawner {
-    /// Try to start the container chain node. In case of error, this panics and stops the node.
-    fn spawn(&self, container_chain_para_id: ParaId) -> impl Future<Output = ()> {
-        let ContainerChainSpawner {
-            orchestrator_chain_interface,
-            orchestrator_client,
-            container_chain_cli,
-            tokio_handle,
-            chain_type,
-            relay_chain,
-            relay_chain_interface,
-            collator_key,
-            sync_keystore,
-            orchestrator_para_id,
-            validator,
-            spawn_handle,
-            spawned_para_ids,
-            collate_on_tanssi: _,
-        } = self.clone();
-        let mut container_chain_cli: ContainerChainCli = container_chain_cli.clone();
-
-        // This closure is used to emulate a try block, it enables using the `?` operator inside
-        let try_closure = move || async move {
-            // Preload genesis data from orchestrator chain storage.
-            // The preload must finish before calling create_configuration, so any async operations
-            // need to be awaited.
-
-            // TODO: the orchestrator chain node may not be fully synced yet,
-            // in that case we will be reading an old state.
-            let orchestrator_chain_info = orchestrator_client.chain_info();
-            log::info!(
-                "Reading container chain genesis data from orchestrator chain at block #{} {}",
-                orchestrator_chain_info.best_number,
-                orchestrator_chain_info.best_hash,
-            );
-            let orchestrator_runtime_api = orchestrator_client.runtime_api();
-
-            log::info!(
-                "Detected assignment for container chain {}",
-                container_chain_para_id
-            );
-
-            let genesis_data = orchestrator_runtime_api
-                .genesis_data(
-                    orchestrator_chain_info.best_hash,
-                    container_chain_para_id.into(),
-                )
-                .expect("error")
-                .ok_or_else(|| {
-                    format!(
-                        "No genesis data registered for container chain id {}",
-                        container_chain_para_id
-                    )
-                })?;
-
-            container_chain_cli
-                .preload_chain_spec_from_genesis_data(
-                    container_chain_para_id.into(),
-                    genesis_data,
-                    chain_type.clone(),
-                    relay_chain.clone(),
-                )
-                .map_err(|e| format!("failed to create container chain chain spec from on chain genesis data: {}", e))?;
-
-            log::info!(
-                "Loaded chain spec for container chain {}",
-                container_chain_para_id
-            );
-
-            // Update CLI params
-            container_chain_cli.base.para_id = Some(container_chain_para_id.into());
-            // New base path
-            /*
-            if container_chain_cli.base.base.tmp {
-                // The temporary folder created when the `--tmp` flag is present is fixed
-                // over the entire process lifetime, so here we replicate the logic to create
-                // a new tmp folder.
-                let temp_dir: tempfile::TempDir = tempfile::Builder::new().prefix("substrate").tempdir()?;
-                let path = PathBuf::from(temp_dir.path());
-                container_chain_cli.base_path = Some(path);
-                container_chain_cli.base.base.tmp = false;
-            }
-            */
-
-            let mut container_chain_cli_config = sc_cli::SubstrateCli::create_configuration(
-                &container_chain_cli,
-                &container_chain_cli,
-                tokio_handle.clone(),
-            )
-            .map_err(|err| format!("Container chain argument error: {}", err))?;
-
-            // This path should be absolute
-            //let base_path = container_chain_cli_config.base_path.as_ref().unwrap().path();
-            //let mut db_path = base_path.clone();
-            //db_path.push("db");
-
-            // The database path is something like
-            // /tmp/zombie-6368e3e070dcee9722a19fb5fa479f21_-4023642-W9VgHHslcbzB/Collator2001-01/data/polkadot/chains/local_testnet/db/full
-            // We want to change the last "db/full" into "db/full-container-{}"
-            let mut db_path = container_chain_cli_config
-                .database
-                .path()
-                .unwrap()
-                .to_owned();
-            log::info!("DB PATH IS {:?}", db_path);
-            db_path.pop();
-            db_path.push(format!("full-container-{}", container_chain_para_id));
-            container_chain_cli_config.database.set_path(&db_path);
-
-            // Start container chain node
-            let (mut container_chain_task_manager, _container_chain_client) =
-                start_node_impl_container(
-                    container_chain_cli_config,
-                    relay_chain_interface.clone(),
-                    orchestrator_chain_interface.clone(),
-                    collator_key.clone(),
-                    sync_keystore.clone(),
-                    container_chain_para_id.into(),
-                    orchestrator_para_id,
-                    validator,
-                )
-                .await?;
-
-            let (signal, on_exit) = exit_future::signal();
-
-            {
-                spawned_para_ids
-                    .lock()
-                    .expect("poison error")
-                    .insert(container_chain_para_id.into(), StopContainerChain(signal));
-            }
-
-            // Emulate task_manager.add_child by making the parent task manager stop if the
-            // container chain task manager stops.
-            // The reverse is also true, if the parent task manager stops, the container chain
-            // task manager will also stop.
-            // But add an additional on_exit future to allow graceful shutdown of container chains.
-            // TODO: not sure what is the difference between spawn and spawn_essential, because
-            // when a "spawn" task panics it stops the node
-            spawn_handle.spawn("container-chain-task-manager", None, async move {
-                let mut t1 = container_chain_task_manager.future().fuse();
-                let mut t2 = on_exit.fuse();
-
-                futures::select! {
-                    res1 = t1 => {
-                        match res1 {
-                            Ok(()) => panic!("container_chain_task_manager has stopped unexpectedly"),
-                            Err(e) => panic!("container_chain_task_manager failed: {}", e),
-                        }
-                    }
-                    _ = t2 => {}
-                }
-            });
-
-            sc_service::error::Result::Ok(())
-        };
-
-        async {
-            match try_closure().await {
-                Ok(()) => {}
-                Err(e) => {
-                    panic!("Failed to start container chain node: {}", e);
-                }
-            }
-        }
-    }
-
-    fn stop(&self, container_chain_para_id: ParaId) {
-        let kill_handle = self
-            .spawned_para_ids
-            .lock()
-            .expect("poison error")
-            .remove(&container_chain_para_id);
-
-        match kill_handle {
-            Some(_signal) => {
-                log::info!("Stopping container chain {}", container_chain_para_id);
-            }
-            None => {
-                log::warn!(
-                    "Tried to stop a container chain that is not running: {}",
-                    container_chain_para_id
-                );
-            }
-        }
-    }
-
-    async fn rx_loop(self, mut rx: tokio::sync::mpsc::UnboundedReceiver<CcSpawnMsg>) {
-        while let Some(msg) = rx.recv().await {
-            match msg {
-                CcSpawnMsg::Spawn(para_id) => {
-                    // For testing
-
-                    if para_id == 2001.into() {
-                        tokio::time::sleep(Duration::from_secs(120)).await;
-                        let tx3 = CCSPAWN
-                            .lock()
-                            .expect("poison error")
-                            .as_ref()
-                            .cloned()
-                            .unwrap();
-                        tx3.send(CcSpawnMsg::Stop(2000.into())).unwrap();
-                    }
-
-                    if para_id == self.orchestrator_para_id {
-                        // Restart collation on orchestrator chain
-                        let f = (self.collate_on_tanssi)();
-                        f.await;
-                    } else {
-                        // Spawn new container chain node
-                        self.spawn_handle.spawn(
-                            "container-chain-spawner",
-                            None,
-                            self.spawn(para_id),
-                        );
-                    }
-                }
-                CcSpawnMsg::Stop(para_id) => {
-                    if para_id == self.orchestrator_para_id {
-                        // Do nothing, because currently the only way to stop collation
-                        // on the orchestrator chain is to start a new container chain.
-                    } else {
-                        self.stop(para_id);
-                    }
-                }
-            }
-        }
-    }
-}
-
 // Log string that will be shown for the container chain: `[Container-2000]`.
 // This needs to be a separate function because the `prefix_logs_with` macro
 // has trouble parsing expressions.
@@ -869,7 +592,7 @@ fn container_log_str(para_id: ParaId) -> String {
 ///
 /// This is the actual implementation that is abstract over the executor and the runtime api.
 #[sc_tracing::logging::prefix_logs_with(container_log_str(para_id))]
-async fn start_node_impl_container(
+pub async fn start_node_impl_container(
     parachain_config: Configuration,
     relay_chain_interface: Arc<dyn RelayChainInterface>,
     orchestrator_chain_interface: Arc<dyn OrchestratorChainInterface>,
